Refactor: replace PsiDocumentableUtils with newly introduced PsiHelper class
- Centralized PSI-related utility logic into the PsiHelper class for better reuse and maintainability.
- Updated DokkaPsiParser and DokkaSymbolVisitor to utilize PsiHelper for bounds resolution, annotation conversion, and checked exception extraction.
- Removed obsolete methods and logic, relocating all relevant functionality to PsiHelper.
